requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-dotenv==1.0.1
opencv-python==4.9.0.80
ultralytics==8.1.2
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
vonage==3.2.0
python-jose==3.3.0
Pillow==10.2.0
numpy==1.26.4
torch==2.0.1
torchvision==0.15.2
bcrypt==4.1.2
flask-limiter==3.5.0
flask-wtf==1.2.1
flask-migrate==4.0.5
flask-bcrypt==1.0.1
flask-mail==0.9.1
flask-jwt-extended==4.6.0
flask-caching==2.1.0
flask-compress==1.14.0
flask-babel==4.0.0
flask-restx==1.3.0
flask-swagger-ui==4.11.1
flask-apscheduler==1.13.1
flask-socketio==5.3.6
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.20.2
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
flower==2.0.1
gunicorn==21.2.0
gevent==23.9.1
eventlet==0.35.2
sentry-sdk==1.40.4
prometheus-client==0.19.0
pytest==8.0.0
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-asyncio==0.23.5
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
bandit==1.7.7
safety==2.3.5 